+++ packages/menu/src/Menu.ts
@@ -150,11 +150,13 @@ export class Menu extends SizedMixin(SpectrumElement, { noDefaultSize: true }) {
     public focusInItemIndex = 0;
 
     /**
-     * whether or not to support pointerdown - drag - pointerup selection strategy
-     * default is true
-     * should be false for mobile to prevent click event being captured behind the menu-tray (cz menu immediately closes on pointerup)
+     * Whether to support the pointerdown-drag-pointerup selection strategy.
+     * Defaults to false to prevent click/touch events from being captured
+     * behind the menu tray in mobile environments (since the menu closes
+     * immediately on pointerup).
      */
-    public shouldSupportDragAndSelect = true;
+
+    public shouldSupportDragAndSelect = false;
 
     public get focusInItem(): MenuItem | undefined {
         return this.rovingTabindexController?.focusInElement;
@@ -391,9 +393,8 @@ export class Menu extends SizedMixin(SpectrumElement, { noDefaultSize: true }) {
                 capture: true,
             }
         );
-
         this.addEventListener('click', this.handleClick);
-        this.addEventListener('mouseover', this.handleMouseover);
+        this.addEventListener('touchend', this.handlePointerup);
         this.addEventListener('focusout', this.handleFocusout);
         this.addEventListener('sp-menu-item-keydown', this.handleKeydown);
         this.addEventListener('pointerup', this.handlePointerup);
@@ -445,17 +446,6 @@ export class Menu extends SizedMixin(SpectrumElement, { noDefaultSize: true }) {
     // if the click and pointerup events are on the same target, we should not
     // handle the click event.
     private pointerUpTarget = null as EventTarget | null;
-
-    private handleMouseover(event: MouseEvent): void {
-        const { target } = event;
-        const menuItem = target as MenuItem;
-        if (
-            this.childItems.includes(menuItem) &&
-            this.isFocusableElement(menuItem)
-        ) {
-            this.rovingTabindexController?.focusOnItem(menuItem);
-        }
-    }
 
     private handleFocusout(): void {
         if (!this.matches(':focus-within'))
@@ -693,7 +683,7 @@ export class Menu extends SizedMixin(SpectrumElement, { noDefaultSize: true }) {
             if (lastFocusedItem?.hasSubmenu) {
                 //open submenu and set focus
                 event.stopPropagation();
-                lastFocusedItem.openOverlay();
+                lastFocusedItem.openOverlay(true);
             }
         } else if (shouldCloseSelfAsSubmenu && this.isSubmenu) {
             event.stopPropagation();
@@ -732,7 +722,7 @@ export class Menu extends SizedMixin(SpectrumElement, { noDefaultSize: true }) {
             // Remove focus while opening overlay from keyboard or the visible focus
             // will slip back to the first item in the menu.
             event.preventDefault();
-            root.openOverlay();
+            root.openOverlay(true);
             return;
         }
         if (key === ' ' || key === 'Enter') {